+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "/home/kevin/anaconda3/envs/scaden/bin/python"
+}
